tsxify StoryPanel
