+++ lib/ReactViews/StandardUserInterface/StandardUserInterface.jsx
@@ -21,9 +21,9 @@ import Notification from "../Notification/Notification";
 import ProgressBar from "../Map/ProgressBar";
 import SidePanel from "../SidePanel/SidePanel";
 import processCustomElements from "./processCustomElements";
-import FullScreenButton from "./../SidePanel/FullScreenButton.jsx";
-import StoryPanel from "./../Story/StoryPanel.jsx";
-import StoryBuilder from "./../Story/StoryBuilder.jsx";
+import FullScreenButton from "./../SidePanel/FullScreenButton";
+import StoryPanel from "./../Story/StoryPanel";
+import StoryBuilder from "./../Story/StoryBuilder";
 
 import PrintView from "../../ReactViews/Map/Panels/SharePanel/Print/PrintView";
