+++ lib/ReactViews/Story/story-panel.scss
@@ -22,10 +22,6 @@
   }
 }
 
-.isPushedUp {
-  bottom: 350px;
-}
-
 .isCentered {
   padding-left: 20px;
   // when changing to full screen, the story panel should move to center slower
